# crates/api/assets/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SoraMetrics API",
  "version": "33.0.0",
  "summary": "Read-only analytics of the SORA v2 blockchain.",
  "description": "Public, read-only, no authentication. Data is indexed from the SORA v2 chain (Substrate) and, for balances, holders, pools, staking and governance, read from a node at request time.\n\nPrices are marginal on-chain quotes (`liquidityProxy.quote` of 0.000001 token to DAI, DAI = 1 USD). `illiquid: true` means selling 1 USD worth of the token returns under 0.50 USD: never derive a market cap or a holding value from that price.\n\nA swap is worth its cheaper priced leg (a marginal quote can overvalue the illiquid side by orders of magnitude). Swap rows carry that value in both `in.usd` and `out.usd`; swap volume, trends, top pair and accumulation are sums of it.\n\nToken supply shown by the site is the official SORA circulating figure, not on-chain total issuance.\n\nRate limits are per IP and per route (10–300 requests per minute); a 429 means wait for the next minute. Agents should prefer the MCP endpoint at `/mcp`, which wraps these routes as typed tools.",
  "license": {
   "name": "MIT",
   "identifier": "MIT"
  }
 },
 "servers": [
  {
   "url": "/",
   "description": "This host"
  }
 ],
...
 "paths": {
...
  "/history/global/liquidity": {
   "get": {
    "operationId": "listLiquidityEvents",
    "summary": "Latest liquidity deposits and withdrawals",
    "description": "",
    "tags": [
     "history"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "description": "Paginated envelope.",
         "properties": {
          "data": {
           "type": "array",
           "items": {
            "type": "object",
            "additionalProperties": true
           }
          },
          "total": {
           "type": "integer"
          },
          "page": {
           "type": "integer"
          },
          "totalPages": {
           "type": "integer"
          }
         },
         "required": [
          "data",
          "total",
          "page",
          "totalPages"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     },
     {
      "name": "limit",
      "in": "query",
      "description": "Rows per page.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 100,
       "default": 25
      }
     },
     {
      "name": "timestamp",
      "in": "query",
      "description": "Upper bound, unix milliseconds: only rows at or before it.",
      "schema": {
       "type": "integer"
      }
     }
    ]
   }
  },
  "/history/global/orderbook": {
   "get": {
    "operationId": "listOrderBookEvents",
    "summary": "Latest order-book events",
    "description": "",
    "tags": [
     "history"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "description": "Paginated envelope.",
         "properties": {
          "data": {
           "type": "array",
           "items": {
            "type": "object",
            "additionalProperties": true
           }
          },
          "total": {
           "type": "integer"
          },
          "page": {
           "type": "integer"
          },
          "totalPages": {
           "type": "integer"
          }
         },
         "required": [
          "data",
          "total",
          "page",
          "totalPages"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     },
     {
      "name": "limit",
      "in": "query",
      "description": "Rows per page.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 100,
       "default": 25
      }
     },
     {
      "name": "type",
      "in": "query",
      "schema": {
       "type": "string",
       "enum": [
        "placed",
        "canceled",
        "executed",
        "filled",
        "market"
       ]
      }
     },
     {
      "name": "timestamp",
      "in": "query",
      "description": "Upper bound, unix milliseconds: only rows at or before it.",
      "schema": {
       "type": "integer"
      }
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
   "Error": {
    "type": "object",
    "properties": {
     "error": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "message": {
      "type": "string"
     }
    },
    "required": [
     "error"
    ]
   }
  }
 }
}